Widget C++ code tidy up
The primary objective of this commit was to tidy up my widget C++ code and comment it where necessary. It did experience issues when binding certain delegates on my widgets but that has been solved thanks to the SynchonizeProperties function. This would have been completed a lot sooner if not for the issues I encountered along the way.

Side note: The Spellbook widget's sliders have tooltips once again, I somehow missed this during the conversion from blueprint to C++.
------------------

New:

SigilUIButtonBasic & SigilUIPercentageSlider

Created these two classes to help clean up the code for the Spellbook widget, it also saves me from having to bind delegates and animations for each button I add to a widget.

SigilTextTooltipWidget

A simple border with a textblock inside, for when I need to display a text tooltip. Currently used for showing the value of the Spellbook widget's sliders.